flask_catgif_clean/requirements.txt
Flask==2.2.2
requests==2.28.1
Werkzeug==2.2.2
mysql-connector-python
